Optimize layout render invalidation with view content hash
